Update build script

+++ scripts/build.sh
@@ -32,6 +32,7 @@ if [ $? -eq 0 ]; then
 
 else
 	echo "FAIL"
+	rm -r $FPIC_DIR
 	exit $?
 
 fi
@@ -45,32 +46,35 @@ if [ $? -eq 0 ]; then
 
 else
 	echo "FAIL"
+	rm -r $FPIC_DIR
 	exit $?
 
 fi
 
 echo "Building and linking recursive fft"
-gcc -c -g -fpic -o $FPIC_DIR/rfft.o $SRC_DIR/recursive_fft.c -lfft -lcomplex -lm
-gcc -shared -o $LIBS_DIR/librfft.so $FPIC_DIR/rfft.o $FPIC_DIR/fft.o $FPIC_DIR/complex.o -lm
+gcc -c -g -fpic -o $FPIC_DIR/rfft.o $SRC_DIR/recursive_fft.c -lfft
+gcc -shared -o $LIBS_DIR/librfft.so $FPIC_DIR/rfft.o $FPIC_DIR/fft.o
 
 if [ $? -eq 0 ]; then
 	echo "OK"
 
 else
 	echo "FAIL"
+	rm -r $FPIC_DIR
 	exit $?
 
 fi
 
 echo "Building and linking iterative fft"
-gcc -c -g -fpic -o $FPIC_DIR/itfft.o $SRC_DIR/iterative_fft.c -lfft -lcomplex -lm
-gcc -shared -o $LIBS_DIR/libitfft.so $FPIC_DIR/itfft.o $FPIC_DIR/fft.o $FPIC_DIR/complex.o -lm
+gcc -c -g -fpic -o $FPIC_DIR/itfft.o $SRC_DIR/iterative_fft.c -lfft
+gcc -shared -o $LIBS_DIR/libitfft.so $FPIC_DIR/itfft.o $FPIC_DIR/fft.o
 
 if [ $? -eq 0 ]; then
 	echo "OK"
 
 else
 	echo "FAIL"
+	rm -r $FPIC_DIR
 	exit $?
 
 fi
@@ -79,31 +83,56 @@ fi
 # gcc -shared -o $LIBS_DIR/libifft.so $FPIC_DIR/ifft.o $FPIC_DIR/fft.o $FPIC_DIR/complex.o -lm
 
 echo "Building and linking conv"
-gcc -c -g -fpic -o $FPIC_DIR/conv.o $SRC_DIR/conv.c -litfft -lfft -lcomplex -lm
-gcc -shared -o $LIBS_DIR/libconv.so $FPIC_DIR/conv.o $FPIC_DIR/itfft.o $FPIC_DIR/fft.o $FPIC_DIR/complex.o -lm
+gcc -c -g -fpic -o $FPIC_DIR/conv.o $SRC_DIR/conv.c -litfft -lfft
+gcc -shared -o $LIBS_DIR/libconv.so $FPIC_DIR/conv.o $FPIC_DIR/itfft.o $FPIC_DIR/fft.o
 
 if [ $? -eq 0 ]; then
 	echo "OK"
 
 else
 	echo "FAIL"
+	rm -r $FPIC_DIR
+	exit $?
+
+fi
+
+echo "Building and linking iterative fft radix 4"
+gcc -c -g -fpic -o $FPIC_DIR/itfft4.o $SRC_DIR/fft_radix_4.c -lfft
+gcc -shared -o $LIBS_DIR/libitfft4.so $FPIC_DIR/itfft4.o $FPIC_DIR/fft.o 
+
+if [ $? -eq 0 ]; then
+	echo "OK"
+
+else
+	echo "FAIL"
+	rm -r $FPIC_DIR
 	exit $?
 
 fi
 
 echo "Building and linking test"
-gcc -g -Wall -Llibs -Wl,-rpath=$LIBS_DIR $SRC_DIR/test.c -lconv -litfft -lfft -lcomplex -lm
-
-# remove the fpic directory
-rm -r $FPIC_DIR
+gcc -g -Wall -Llibs -Wl,-rpath=$LIBS_DIR $SRC_DIR/test.c -lconv -litfft4 -litfft -lfft -lcomplex -lm
 
 if [ $? -eq 0 ]; then
 	echo "OK"
 
 else
 	echo "FAIL"
+	rm -r $FPIC_DIR
 	exit $?
 
 fi
+
+# remove the fpic directory
+rm -r $FPIC_DIR
+
+# if [ $? -eq 0 ]; then
+# 	echo "OK"
+
+# else
+# 	echo "FAIL"
+# 	exit $?
+
+# fi
 
 echo "Finished with 0 errors"
